+++ IsekaiMod/Content/Features/Deathsnatcher/DeathsnatcherCreateUndead.cs
@@ -62,8 +62,6 @@ namespace IsekaiMod.Content.Features.Deathsnatcher
                 bp.AddComponent<ContextRankConfig>(c => {
                     c.m_Type = AbilityRankType.Default;
                     c.m_BaseValueType = ContextRankBaseValueType.ClassLevel;
-                    c.m_Stat = StatType.Unknown;
-                    c.m_SpecificModifier = ModifierDescriptor.None;
                     c.m_Progression = ContextRankProgression.BonusValue;
                     c.m_StepLevel = 1;
                     c.m_Class = new BlueprintCharacterClassReference[] { DeathsnatcherClass.GetReference() };